+++ changelog.md
@@ -1,14 +1,19 @@
 # Change Log
 
 @TODO:
-- Add external timelag file as input.
+
+- Simplify function arguments using OOP.
 - Replace Theil-Sen estimator with a more efficient robust linear regression method (perhaps MM-estimator?).
 - Use `pandas.Timestamp` as the standard timestamp passed between functions.
+
+# 0.2.0
+
 
 # 0.1.12 (2017-04-24)
 
 - Added suggestive quality flags based on the internal differences among the flux values calculated from three fitting methods.
 - Added options to save configuration files in the output folder.
+- Implemented a new timelag method: `'prescribed'`. This method uses an external timelag data file to determine the timelag value.
 
 # 0.1.11 (2017-04-23)